add unit tests for ArraySortes and fix arrays type

+++ calc.Tests/HelpersTests/ArrayHelpersTest.cs
@@ -7,19 +7,19 @@ namespace calc.Helpers
     [TestFixture]
     public  class ArrayHelpersTest
     {
-        [TestCase("1 2 3 4 5", ' ', new int[] { 1, 2, 3, 4, 5 })]
-        [TestCase("-5", ' ', new int[] { -5 })]
+        [TestCase("1 2 3 4 5", ' ', new double[] { 1, 2, 3, 4, 5 })]
+        [TestCase("-5", ' ', new double[] { -5 })]
 
-        public void StringToArrayTest(string toSplit,char delimiter, int[] result)
+        public void StringToArrayTest(string toSplit,char delimiter, double[] result)
         {
 
             var testResult = ArrayHelpers.StringToArray(toSplit, delimiter);
             Assert.AreEqual(testResult, result);
         }
 
-        [TestCase(new int[] {1, 2, 3, 4, 5}, ',', "1,2,3,4,5")]
-        [TestCase(new int[] { -5 }, ' ', "-5")]
-        public void ArrayToStringTest(int[] array, char delimiter, string result)
+        [TestCase(new double[] {1, 2, 3, 4, 5}, ',', "1,2,3,4,5")]
+        [TestCase(new double[] { -5 }, ' ', "-5")]
+        public void ArrayToStringTest(double[] array, char delimiter, string result)
         {
             var testResult = ArrayHelpers.ArrayToString(array, delimiter);
             Assert.AreEqual(testResult, result);
@@ -37,7 +37,7 @@ namespace calc.Helpers
         [ExpectedException(typeof(Exception))]
         public void EmptyArrayToStringTest()
         {
-            ArrayHelpers.ArrayToString(new int[]{}, ',');
+            ArrayHelpers.ArrayToString(new double[]{}, ',');
         }
 
         [Test]
